refactor(delivery): extract AcceptanceTestConfig class with common logic for acceptance tests
